+++ CHANGELOG.md
@@ -31,6 +31,7 @@
  * Added ReportedFailedException which causes a batch to fail without killing worker and reports the error to the UI
  * Execute latency now tracked and shown in Storm UI
  * Adding testTuple methods for easily creating Tuple instances to Testing API (thanks xumingming)
+ * Trident now throws an error during construction of a topology when try to select fields that don't exist in a stream (thanks xumingming)
  * Bug fix: Fix race condition in supervisor that would lead to supervisor continuously crashing due to not finding "stormconf.ser" file for an already killed topology
  * Bug fix: bin/storm script now displays a helpful error message when an invalid command is specified